requirements.txt
fastapi==0.116.1
protobuf==6.31.1
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.46.1
uvicorn==0.35.0
google-genai==1.25.0